refactor(frontend): split match editor tabs
